Hello Wes!

Gatsby maintainer here -- hope you have a great day, and this PR will also bring you some joy! :)

We've been making some improvements to "gatsby develop" starting with 2.26 (interested parties can read it here: https://www.gatsbyjs.com/docs/reference/release-notes/). After some first experiments and partial rollouts we now finally shipped those improvements in 2.30 for everyone: https://www.gatsbyjs.com/docs/reference/release-notes/v2.30/

Most importantly "Lazy images" & "Query on Demand". While we also made improvements to different plugins that you use, those two things impacted your "develop (cold)" the most. What I mean by that:
Starting off with a cold cache, so as if one would clone the repo, install deps and run `gatsby develop`.

## Update (2020-01-22)

Since my initial comment here the 2.31 version was released and thus I updated the PR to use stable versions of packages now.

I also removed the `root` option from the MDX plugin as digging into the code a bit it turned out it's a noop.

---

## How `gatsby develop` changes

The index page and its images will be created as usual, but any subsequent pages (and their queries + images) are processed on demand. This already happens when you hover over a link. So the whole upfront generation of pages & images is gone! :)

In case you have a long running query, you'll see a loading indicator:

![](https://user-images.githubusercontent.com/419821/102150822-bee77b00-3e71-11eb-8a52-e02db0b8aa43.gif)

## Why is this PR in draft mode?

Good question! I wanted to test out some changes that are currently only in our `next` version, so not in stable yet. These changes will be released with 2.31 on 19th of January. Until then I'll also see to fix the issue that `gatsby-plugin-mdx` isn't recognizing the `root` option.

Lastly, I haven't done much testing yet if everything still works as it should 😅 

## Awesome if you could do...

Check it out and see if the site behaves normally and if the result of `gatsby build` is also improved. I haven't looked into it yet but it should improve, too 👍 

## Comparison

Comparison on my machine (only one run, so there will be some variance in your results) with Ryzen 9 3900X, 32GB RAM and WSL2:

<details>
 <summary>Before/After Logs</summary>

**Before:**

```shell
$ gatsby develop
success open and validate gatsby-configs - 0.061s
success load plugins - 2.479s
success onPreInit - 0.036s
success initialize cache - 0.007s
success copy gatsby files - 0.335s
success onPreBootstrap - 0.033s
success createSchemaCustomization - 0.291s
success Checking for changed pages - 0.003s
success source and transform nodes - 22.081s
success building schema - 1.115s
info Total nodes: 2935, SitePage nodes: 411 (use --verbose for breakdown)
success createPages - 28.401s
success Checking for changed pages - 0.002s
success createPagesStatefully - 0.187s
success update schema - 0.046s
success write out redirect data - 0.006s
success onPostBootstrap - 0.002s
info bootstrap finished - 59.403s
success onPreExtractQueries - 0.003s
success extract queries from components - 0.707s
success write out requires - 0.078s
success run static queries - 0.229s - 2/2 8.75/s
warn You can't use childImageSharp together with linebubble.mov.gif — use publicURL instead. The childImageSharp portion of the query in this file will return
null:
warn You can't use childImageSharp together with linebubble.mov.gif — use publicURL instead. The childImageSharp portion of the query in this file will return
null:
success run page queries - 32.461s - 423/423 13.03/s
success Generating image thumbnails - 627.577s - 1936/1936 3.08/s
```

**After:**

```
$ gatsby develop
success open and validate gatsby-configs - 0.093s
warn Warning: there are unknown plugin options for "gatsby-plugin-mdx": root
Please open an issue at ghub.io/gatsby-plugin-mdx if you believe this option is valid.
success load plugins - 2.618s
success onPreInit - 0.034s
success initialize cache - 0.006s
success copy gatsby files - 0.510s
success onPreBootstrap - 0.027s
success createSchemaCustomization - 0.274s
success Checking for changed pages - 0.001s
success source and transform nodes - 15.222s
success building schema - 0.409s
info Total nodes: 2935, SitePage nodes: 411 (use --verbose for breakdown)
success createPages - 22.716s
success Checking for changed pages - 0.001s
success createPagesStatefully - 0.107s
success update schema - 0.037s
success write out redirect data - 0.005s
success onPostBootstrap - 0.002s
info bootstrap finished - 45.592s
success onPreExtractQueries - 0.001s
success extract queries from components - 0.845s
success write out requires - 0.084s
success run static queries - 0.180s - 2/2 11.13/s
success run page queries - 0.011s - 2/2 181.62/s
```
</details>

**So from roughly 774 seconds to roughly 85 seconds is a 89% improvement!**