@LekoArts 2021-02-24
Hi @wesbos 👋 

Was this PR superseded by your merch store changes? :) If yes, then I can close this of course! Hope you like the performance boost so far.