@wesbos 2021-02-24
@LekoArts Yes! Sorry but I did upgrade Gatsby, It's nice a quick to start up again! I didn't grab any stats, but it's definitely noticable!